+++ Engine/Module/ModuleScene.h
@@ -23,7 +23,7 @@ public:
 	void RemoveGameObject(GameObject * game_object_to_remove);
 
 	GameObject* GetRoot() const;
-	GameObject* GetGameObject(unsigned int UUID) const;
+	GameObject* GetGameObject(uint64_t UUID) const;
 
 	void DeleteCurrentScene();
 	void Save(Config& serialized_scene) const;
